corrected mocha tests structure
